Move custom macros to their own file.
